Add Distributed.jl support
When used with `Distributed.jl`, `DirWalker`s can be created with
`RemoteChannel`s by passing `RemoteChannel` to the `DirWalker` function.
In this case, a `RemoteDirWalker` is returned, which is a subtype of
`AbstractDirWalker`.

It is important that the file predicate function `filepred` and the file
transforming function `filefunc` are defined on all worker processes
(e.g. with `@everywhere`).