Major refactoring, queues now separate
The DirWalker struct has been eliminated.  This allows for the directory
queue to be a Channel and the file and out queues to be RemoteChannels.

To simplify the package Distibuted is now a direct dependency rather
than an extension.  The primary motivation for that is to be able to
provide RemoteDirQueue, RemoteFileQueue, and RemoteOutQueue types, which
can't be exposed when defined only in an extension.

The results from the dagents and fagents are now fetched before
returning from start_dirwalker.