Rename start_dirwalker to run_dirwalker
Renamed start_dirwalker to run_dirwalker and make `run_dirwalker` run
synchronously in the calling Task.  This give the caller the choice of
how/whether to run the walker synchronously or asynchronously (e.g. on
a remote worker or in a new Task).  The old way always ran the walker in
a Task in the same process, but this new way allows more options (i.e.
synchronously in the calling Task, async in a new Task of the same
process, or async on a remote worker).