Expect file function to return an iterator
Instead of expecting a single object (or a single Vector of objects) to
be returned from the file function, _process_files now expects an
iterator that yields objects (or even a single object).  This allows the
file function to partition large objects (e.g. large Vectors of objects)
into smaller pieces to avoid excessive memory consumption in the output
queue.  File functions that used to return long Vectors of objects can
now partition the vector into smaller subsets.  It also allows the file
function to return a partitioned lazy map to avoid having to materialize
the long list in the worker process, thereby reducing memory pressure
there as well.